feat:使用环形缓冲区替代serial中的接收数组，添加获取serial数据的方式

+++ User/Tools/RingBuffer/ring_buf.c
@@ -66,7 +66,7 @@ char RingBufPop(RingBuf* ring)
     return byte;
 }
 
-uint8_t RingBufRead(RingBuf* ring, char* buf, int len)
+uint32_t RingBufRead(RingBuf* ring, char* buf, int len)
 {
     if (ring == NULL || ring->unit == NULL || ring->unit->start == NULL || buf == NULL || len <= 0) {
         return 0;
